chakula_connect/backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
pandas
scikit-learn
africastalking
# sentinelhub-py
python-dotenv
django-cors-headers
whitenoise
djangorestframework-simplejwt
